Some fixes to TVP parameter property discovery
